# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo",
    "description": "The Igloo control plane API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Gets a job.",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}/logs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Streams a job's output as server-sent events: `stdout` and `stderr` events whose id resumes\nthe stream through `Last-Event-ID`, then one `end` event with the finished job.",
        "operationId": "getJobLogs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "description": "Resume after this event",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "Lists repositories, oldest first.",
        "operationId": "listRepos",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepoResource"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "repos"
        ],
        "summary": "Registers a repository.",
        "operationId": "registerRepo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRepoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoResource"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos/{id}": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "Gets a repository.",
        "operationId": "getRepo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/repos/{id}/secrets": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "Lists the names of a repository's secrets; values are never returned.",
        "operationId": "listRepoSecrets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretList"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos/{id}/secrets/{name}": {
      "put": {
        "tags": [
          "repos"
        ],
        "summary": "Sets or replaces a secret. The body is its value, as text.",
        "operationId": "setRepoSecret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "`[A-Z_][A-Z0-9_]*`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {},
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "repos"
        ],
        "summary": "Deletes a secret.",
        "operationId": "deleteRepoSecret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {},
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repos/{id}/snapshots": {
      "post": {
        "tags": [
          "repos"
        ],
        "summary": "Snapshots a repository at a commit: the checkout under `/workspace`, with a shallow `.git`,\nover an optional base snapshot. A branch is fetched from the forge first.",
        "operationId": "snapshotRepo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepoSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoSnapshotResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "description": "One invalid field of a request.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field path, such as `\"limits.millicpus\"`."
          },
          "message": {
            "type": "string",
            "description": "Why it is invalid."
          }
        }
      },
...
      "JobPhase": {
        "type": "string",
        "description": "Where a job is in its lifecycle.",
        "enum": [
          "queued",
          "leased",
          "running",
          "finished",
          "failed",
          "cancelled"
        ]
      },
      "JobResource": {
        "type": "object",
        "description": "A job.",
        "required": [
          "id",
          "sandbox",
          "argv",
          "env",
          "timeout_seconds",
          "phase",
          "submitted_at"
        ],
        "properties": {
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The program and its arguments."
          },
          "env": {
            "type": "object",
            "description": "Extra environment.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "exit_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The exit code, when finished."
          },
          "failure_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why it failed, when failed."
          },
          "id": {
            "type": "string",
            "description": "Its id (`job_...`)."
          },
          "phase": {
            "$ref": "#/components/schemas/JobPhase",
            "description": "The phase."
          },
          "sandbox": {
            "type": "string",
            "description": "The sandbox it runs in."
          },
          "secrets": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The repository secrets it is given, by name."
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was submitted."
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "How long the process may run.",
            "minimum": 0
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "An error response. `code` is stable and machine-readable; `title` and `detail` are for\nhumans.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The stable error code, such as `\"sandbox.not_found\"`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "What went wrong in this occurrence."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Every invalid field, for validation problems."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          }
        }
      },
      "RegisterRepoRequest": {
        "type": "object",
        "description": "Registers a repository.",
        "required": [
          "location"
        ],
        "properties": {
          "default_branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "The branch changes merge into; `main` when omitted."
          },
          "location": {
            "type": "string",
            "description": "`github.com/<owner>/<name>`, or an absolute path on the server for development."
          },
          "token_secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "The repository secret holding the forge token, such as `GITHUB_TOKEN`."
          }
        }
      },
      "RepoResource": {
        "type": "object",
        "description": "A repository.",
        "required": [
          "id",
          "location",
          "default_branch"
        ],
        "properties": {
          "default_branch": {
            "type": "string",
            "description": "The branch changes merge into."
          },
          "id": {
            "type": "string",
            "description": "Its id (`repo_...`)."
          },
          "location": {
            "type": "string",
            "description": "Where it is hosted."
          },
          "token_secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "The secret holding the forge token."
          }
        }
      },
      "RepoSnapshotRequest": {
        "type": "object",
        "description": "Makes a snapshot of a repository at a commit: its checkout under `/workspace`, with a shallow\n`.git`, over `base`. Names exactly one of `branch` (fetched first) and `commit`.",
        "properties": {
          "base": {
            "type": [
              "string",
              "null"
            ],
            "description": "The snapshot to check out over, such as an imported image."
          },
          "branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "A branch on the forge."
          },
          "commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "A commit of the default branch's history."
          }
        }
      },
      "RepoSnapshotResource": {
        "type": "object",
        "description": "A repository snapshot.",
        "required": [
          "snapshot",
          "commit"
        ],
        "properties": {
          "commit": {
            "type": "string",
            "description": "The commit checked out."
          },
          "snapshot": {
            "type": "string",
            "description": "The snapshot."
          }
        }
      },
...
      "SecretList": {
        "type": "object",
        "description": "The names of a repository's secrets. Values are never returned.",
        "required": [
          "names"
        ],
        "properties": {
          "names": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The names, sorted."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}